move files related to map-check
