+++ renderer/src/web/map-check/MapStatButton.vue
@@ -17,7 +17,7 @@ import { computed, useCssModule } from 'vue'
 import { AppConfig } from '@/web/Config'
 import { PreparedStat } from './prepare-map-stats'
 import { ItemCheckWidget } from '../overlay/interfaces'
-import { nextDecision, decisionCreate } from '../settings/maps/common.js'
+import { nextDecision, decisionCreate } from './common.js'
 
 import ItemModifierText from '../ui/ItemModifierText.vue'
