+++ renderer/src/web/map-check/SettingsMatcherEntry.vue
