+++ renderer/src/web/settings/SettingsWindow.vue
@@ -63,7 +63,7 @@ import SettingsAbout from './about.vue'
 import SettingsPricecheck from './price-check.vue'
 import SettingsItemcheck from './item-check.vue'
 import SettingsDebug from './debug.vue'
-import SettingsMaps from './maps/maps.vue'
+import SettingsMaps from '../map-check/settings-maps.vue'
 import SettingsStashSearch from './stash-search.vue'
 import SettingsStopwatch from './stopwatch.vue'
 import SettingsItemSearch from './item-search.vue'
